added set price when user clicks on list

+++ lib/providers/data_provider.dart
@@ -1,33 +1,33 @@
 import 'package:flutter/material.dart';
 
 class DataProvider with ChangeNotifier {
-  int? _distanta;
-  int? _consum;
-  int? _pret;
+  double? _distanta;
+  double? _consum;
+  double? _pret;
 
-  int? get distanta {
+  double? get distanta {
     return _distanta;
   }
 
-  int? get consum {
+  double? get consum {
     return _consum;
   }
 
-  int? get pret {
+  double? get pret {
     return _pret;
   }
 
-  void setDistanta(int distanta) {
+  void setDistanta(double distanta) {
     _distanta = distanta;
     notifyListeners();
   }
 
-  void setConsum(int consum) {
+  void setConsum(double consum) {
     _consum = consum;
     notifyListeners();
   }
 
-  void setPret(int pret) {
+  void setPret(double pret) {
     _pret = pret;
     notifyListeners();
   }
